+++ apps/frontend/src/modules/quiz/quiz-api.ts
@@ -5,8 +5,8 @@ import {
 	type QuizAnswersRequestDto,
 	type QuizQuestionDto,
 	type QuizScoresGetAllResponseDto,
-	type QuizScoresResponseDto,
 	type QuizScoresUpdateRequestDto,
+	type QuizScoresUpdateResponseDto,
 	type QuizUserAnswerDto,
 } from "~/modules/quiz/quiz.js";
 
@@ -19,7 +19,7 @@ class QuizApi extends BaseHTTPApi {
 
 	public async editScores(
 		payload: QuizScoresUpdateRequestDto,
-	): Promise<QuizScoresResponseDto> {
+	): Promise<QuizScoresUpdateResponseDto> {
 		const response = await this.load(
 			this.getFullEndpoint(QuizApiPath.SCORE, {}),
 			{
@@ -30,7 +30,7 @@ class QuizApi extends BaseHTTPApi {
 			},
 		);
 
-		return await response.json<QuizScoresResponseDto>();
+		return await response.json<QuizScoresUpdateResponseDto>();
 	}
 
 	public async getAllQuestions(): Promise<{ items: QuizQuestionDto[][] }> {