+++ apps/frontend/src/modules/tasks/libs/types/types.ts
@@ -1,9 +1,10 @@
+export { type AddTaskNoteHandler } from "./add-task-note-handler.type.js";
+export { type TaskNoteRequestDto } from "./task-note-request-dto.type.js";
 export { type TaskUpdatePayload } from "./task-update-payload.type.js";
 export {
 	type TaskCreateDto,
 	type TaskDto,
 	type TaskNoteDto,
 	type TaskNoteParametersDto,
-	type TaskNoteRequestDto,
 	type TaskUpdateRequestDto,
 } from "shared";