+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.28.0](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/compare/shared-v1.27.0...shared-v1.28.0) (2024-09-24)
+
+
+### Features
+
+* **backend/frontend/shared:** Add notes to tasks bb-419 ([#535](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/issues/535)) ([c8ed822](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/commit/c8ed822f1f3333d81888f0c6e7467b87ca989dd4))
+* **frontend/backend/shared:** Add date on the "My wheel results" page bb-494 ([#557](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/issues/557)) ([5b9a41f](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/commit/5b9a41f831f1aad034cccb02991d71894f836db5))
+* **shared:** requirements for authorization pages for web version bb-209 ([#552](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/issues/552)) ([29c7631](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/commit/29c7631d19c665bb05bcc23c4717cb2054b85db7))
+
 ## [1.27.0](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/compare/shared-v1.26.0...shared-v1.27.0) (2024-09-23)